# bff-bank123-openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/bff-bank123/usuario/v1/perfil":{"get":{"tags":["usuario-controller"],"operationId":"obterPerfil","parameters":[{"name":"x-account-id","in":"header","required":true,"schema":{"type":"integer","format":"int32"},"description":"ID da conta do usuário"},{"name":"x-correlationId","in":"header","required":false,"schema":{"type":"string","default":"uuid-fake"},"description":"ID de correlação para rastreamento de requisições"},{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"},"description":"Token de autorização JWT"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PerfilResponse"}}}}}}},"/bff-bank123/extrato/v1/saldo":{"get":{"tags":["extrato-controller"],"operationId":"consultarSaldo","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}},{"name":"x-account-id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"x-correlationId","in":"header","required":false,"schema":{"type":"string","default":"uuid-fake"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SaldoResponse"}}}}}}},"/bff-bank123/extrato/v1/listagem":{"get":{"tags":["extrato-controller"],"operationId":"listarExtrato","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}},{"name":"x-account-id","in":"header","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"x-correlationId","in":"header","required":false,"schema":{"type":"string","default":"uuid-fake"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransacaoResponse"}}}}}}}}},"components":{"schemas":{"PerfilResponse":{"type":"object","properties":{"nomeCompleto":{"type":"string"},"endereco":{"type":"string"},"agencia":{"type":"string"},"numeroConta":{"type":"integer","format":"int32"}}},"SaldoResponse":{"type":"object","properties":{"numeroConta":{"type":"integer","format":"int32"},"saldo":{"type":"number"}}},"TransacaoResponse":{"type":"object","properties":{"idTransacao":{"type":"integer","format":"int32"},"dataTransacao":{"type":"string","format":"date-time"},"valorTransacao":{"type":"number"},"operacao":{"type":"string"},"destino":{"type":"string"},"origem":{"type":"string"}}}}}}